Split batch copy into separate classes
