+++ src/execution/physical_plan/plan_copy_to_file.cpp
@@ -1,6 +1,7 @@
 #include "duckdb/execution/physical_plan_generator.hpp"
 #include "duckdb/execution/operator/persistent/physical_copy_to_file.hpp"
 #include "duckdb/execution/operator/persistent/physical_batch_copy_to_file.hpp"
+#include "duckdb/execution/operator/persistent/physical_fixed_batch_copy.hpp"
 #include "duckdb/planner/operator/logical_copy_to_file.hpp"
 
 namespace duckdb {
@@ -28,12 +29,21 @@ unique_ptr<PhysicalOperator> PhysicalPlanGenerator::CreatePlan(LogicalCopyToFile
 			throw InternalException("BATCH_COPY_TO_FILE can only be used if batch indexes are supported");
 		}
 		// batched copy to file
-		auto copy = make_uniq<PhysicalBatchCopyToFile>(op.types, op.function, std::move(op.bind_data),
-		                                               op.estimated_cardinality);
-		copy->file_path = op.file_path;
-		copy->use_tmp_file = op.use_tmp_file;
-		copy->children.push_back(std::move(plan));
-		return std::move(copy);
+		if (op.function.desired_batch_size) {
+			auto copy = make_uniq<PhysicalFixedBatchCopy>(op.types, op.function, std::move(op.bind_data),
+			                                              op.estimated_cardinality);
+			copy->file_path = op.file_path;
+			copy->use_tmp_file = op.use_tmp_file;
+			copy->children.push_back(std::move(plan));
+			return std::move(copy);
+		} else {
+			auto copy = make_uniq<PhysicalBatchCopyToFile>(op.types, op.function, std::move(op.bind_data),
+			                                               op.estimated_cardinality);
+			copy->file_path = op.file_path;
+			copy->use_tmp_file = op.use_tmp_file;
+			copy->children.push_back(std::move(plan));
+			return std::move(copy);
+		}
 	}
 	// COPY from select statement to file
 	auto copy = make_uniq<PhysicalCopyToFile>(op.types, op.function, std::move(op.bind_data), op.estimated_cardinality);
